+++ v2/CHANGELOG.md
@@ -6,6 +6,7 @@
 - Adjust Cursor options
 - Switch to Go 1.22.8
 - Remove deprecated context functions
+- Fix Error Handler in CreateCollectionWithOptions
 
 ## [2.1.1](https://github.com/arangodb/go-driver/tree/v2.1.1) (2024-09-27)
 - Improve backup tests stability